requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.8
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jira==2.0.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysql-connector-python==8.0.17
numpy==1.17.2
oauthlib==3.1.0
parso==0.5.1
pbr==5.4.3
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pywin32==225
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
wxPython==4.0.6
